% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cranlogs-package.R
\docType{package}
\name{cranlogs-package}
\alias{cranlogs}
\alias{cranlogs-package}
\title{cranlogs: Download Logs from the 'RStudio' 'CRAN' Mirror}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

'API' to the database of 'CRAN' package downloads from the
    'RStudio' 'CRAN mirror'. The database itself is at <http://cranlogs.r-pkg.org>,
    see <https://github.com/r-hub/cranlogs.app> for the raw 'API'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-hub/cranlogs}
  \item \url{https://r-hub.github.io/cranlogs}
  \item Report bugs at \url{https://github.com/r-hub/cranlogs/issues}
}

}
\author{
\strong{Maintainer}: Gábor Csárdi \email{csardi.gabor@gmail.com}

Other contributors:
\itemize{
  \item R Consortium [funder]
}

}
\keyword{internal}
