% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{cranlogs_badge}
\alias{cranlogs_badge}
\title{Create Markdown code for a cranlogs badge}
\usage{
cranlogs_badge(package_name, summary = c("last-month", "last-day",
  "last-week", "grand-total"), color = "blue")
}
\arguments{
\item{package_name}{name of the package}

\item{summary}{One of "last-day", "last-week", "last-month", "grand-total".}

\item{color}{Either a hex color code or one of "brightgreen", "green", "yellowgreen",
"yellow", "orange", "red", "lightgrey", "blue".}
}
\value{
The string to be copy-pasted to the README source.
}
\description{
Create Markdown code for a cranlogs badge
}
\examples{
cranlogs_badge("praise")
cranlogs_badge("praise", summary = "grand-total")
cranlogs_badge("praise", summary = "last-week", color = "orange")
cranlogs_badge("praise", color = "ff8c69")
}
