% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_parallel.r
\name{message_parallel}
\alias{message_parallel}
\title{Message parallel}
\usage{
message_parallel(...)
}
\value{
Null
}
\description{
Send messages to console even from within parallel processes
}
\keyword{internal}
